fix(build): avoid circular dependency in icons setup
